chore: refactored flow code
