+++ src/commands/flow/internal/isSessionStateError.ts
@@ -1,0 +1,2 @@
+export const isSessionStateError = (err: string) =>
+  err.includes('Could not get session state')
