+++ src/commands/flow/internal/normalizeFilePath.ts
@@ -1,0 +1,8 @@
+import { getRealPath } from '@sasjs/utils'
+import path from 'path'
+
+export const normalizeFilePath = (filePath: string) => {
+  const pathSepRegExp = new RegExp(path.sep.replace(/\\/g, '\\\\'), 'g')
+
+  return getRealPath(filePath).replace(pathSepRegExp, '/')
+}
